Update README.md
# Loan Approval Classifier using Perceptron Learning

This project implements a simple binary classifier using the Perceptron Learning Algorithm to predict loan approval outcomes based on applicant data.

## 📌 Project Overview

The goal of this project is to classify whether a loan should be approved or not based on features like Existing loan and Credit Score. The model is trained using a custom implementation of the perceptron algorithm.

## 🔧 Technologies & Tools

- Python
- NumPy
- Pandas (for data handling)
- Matplotlib (for visualization)

## 🧠 Algorithm

The perceptron is a basic unit of a neural network. It is a linear classifier that updates weights based on prediction errors. The update rule is:

w = w + (xi*yi)
w_0 = w_0 + yi